requirements.txt
python-dateutil==2.8.2
types-python-dateutil-2.8.0
python-dotenv==0.19.0
requests==2.26.0
sqlmodel==0.0.4
types-python-dateutil==2.8.0
psycopg2==2.9.1
